split off EthFilterService
