+++ src/main/scala/io/iohk/ethereum/jsonrpc/EthService.scala
@@ -85,30 +85,41 @@ object EthService {
   case class IeleCallResponse(returnData: Seq[ByteString])
   case class EstimateGasResponse(gas: BigInt)
 
-  case class NewFilterRequest(filter: Filter)
-  case class Filter(
-      fromBlock: Option[BlockParam],
-      toBlock: Option[BlockParam],
-      address: Option[Address],
-      topics: Seq[Seq[ByteString]]
-  )
+  // case class NewFilterRequest(filter: Filter)
+  // case class Filter(
+  //     fromBlock: Option[BlockParam],
+  //     toBlock: Option[BlockParam],
+  //     address: Option[Address],
+  //     topics: Seq[Seq[ByteString]]
+  // )
 
-  case class NewBlockFilterRequest()
-  case class NewPendingTransactionFilterRequest()
+  // case class NewBlockFilterRequest()
+  // case class NewPendingTransactionFilterRequest()
 
-  case class NewFilterResponse(filterId: BigInt)
+  // case class NewFilterResponse(filterId: BigInt)
 
-  case class UninstallFilterRequest(filterId: BigInt)
-  case class UninstallFilterResponse(success: Boolean)
+  // case class UninstallFilterRequest(filterId: BigInt)
+  // case class UninstallFilterResponse(success: Boolean)
 
-  case class GetFilterChangesRequest(filterId: BigInt)
-  case class GetFilterChangesResponse(filterChanges: FilterChanges)
+  // case class GetFilterChangesRequest(filterId: BigInt)
+  // case class GetFilterChangesResponse(filterChanges: FilterChanges)
 
-  case class GetFilterLogsRequest(filterId: BigInt)
-  case class GetFilterLogsResponse(filterLogs: FilterLogs)
+  // case class GetFilterLogsRequest(filterId: BigInt)
+  // case class GetFilterLogsResponse(filterLogs: FilterLogs)
 
-  case class GetLogsRequest(filter: Filter)
-  case class GetLogsResponse(filterLogs: LogFilterLogs)
+  // case class GetLogsRequest(filter: Filter)
+  // case class GetLogsResponse(filterLogs: LogFilterLogs)
+  // case class GetUncleCountByBlockNumberRequest(block: BlockParam)
+  // case class GetUncleCountByBlockNumberResponse(result: BigInt)
+
+  // case class GetUncleCountByBlockHashRequest(blockHash: ByteString)
+  // case class GetUncleCountByBlockHashResponse(result: BigInt)
+
+  // case class GetCoinbaseRequest()
+  // case class GetCoinbaseResponse(address: Address)
+
+  // case class GetBlockTransactionCountByNumberRequest(block: BlockParam)
+  // case class GetBlockTransactionCountByNumberResponse(result: BigInt)
 
   case class ResolvedBlock(block: Block, pendingState: Option[InMemoryWorldStateProxy])
 
@@ -249,72 +260,105 @@ class EthService(
     }
   }
 
-  def newFilter(req: NewFilterRequest): ServiceResponse[NewFilterResponse] = {
-    implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
+  // def newFilter(req: NewFilterRequest): ServiceResponse[NewFilterResponse] = {
+  //   implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
 
-    import req.filter._
-    filterManager
-      .askFor[FM.NewFilterResponse](FM.NewLogFilter(fromBlock, toBlock, address, topics))
-      .map { resp =>
-        Right(NewFilterResponse(resp.id))
-      }
-  }
+  //   import req.filter._
+  //   filterManager
+  //     .askFor[FM.NewFilterResponse](FM.NewLogFilter(fromBlock, toBlock, address, topics))
+  //     .map { resp =>
+  //       Right(NewFilterResponse(resp.id))
+  //     }
+  // }
 
-  def newBlockFilter(req: NewBlockFilterRequest): ServiceResponse[NewFilterResponse] = {
-    implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
-    filterManager
-      .askFor[FM.NewFilterResponse](FM.NewBlockFilter)
-      .map { resp =>
-        Right(NewFilterResponse(resp.id))
-      }
-  }
+  // def newBlockFilter(req: NewBlockFilterRequest): ServiceResponse[NewFilterResponse] = {
+  //   implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
+  //   filterManager
+  //     .askFor[FM.NewFilterResponse](FM.NewBlockFilter)
+  //     .map { resp =>
+  //       Right(NewFilterResponse(resp.id))
+  //     }
+  // }
 
-  def newPendingTransactionFilter(req: NewPendingTransactionFilterRequest): ServiceResponse[NewFilterResponse] = {
-    implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
-    filterManager
-      .askFor[FM.NewFilterResponse](FM.NewPendingTransactionFilter)
-      .map { resp =>
-        Right(NewFilterResponse(resp.id))
-      }
-  }
+  // def newPendingTransactionFilter(req: NewPendingTransactionFilterRequest): ServiceResponse[NewFilterResponse] = {
+  //   implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
+  //   filterManager
+  //     .askFor[FM.NewFilterResponse](FM.NewPendingTransactionFilter)
+  //     .map { resp =>
+  //       Right(NewFilterResponse(resp.id))
+  //     }
+  // }
 
-  def uninstallFilter(req: UninstallFilterRequest): ServiceResponse[UninstallFilterResponse] = {
-    implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
+  // def uninstallFilter(req: UninstallFilterRequest): ServiceResponse[UninstallFilterResponse] = {
+  //   implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
 
-    filterManager
-      .askFor[FM.UninstallFilterResponse.type](FM.UninstallFilter(req.filterId))
-      .map(_ => Right(UninstallFilterResponse(success = true)))
-  }
+  //   filterManager
+  //     .askFor[FM.UninstallFilterResponse.type](FM.UninstallFilter(req.filterId))
+  //     .map(_ => Right(UninstallFilterResponse(success = true)))
+  // }
 
-  def getFilterChanges(req: GetFilterChangesRequest): ServiceResponse[GetFilterChangesResponse] = {
-    implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
+  // def getFilterChanges(req: GetFilterChangesRequest): ServiceResponse[GetFilterChangesResponse] = {
+  //   implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
 
-    filterManager
-      .askFor[FM.FilterChanges](FM.GetFilterChanges(req.filterId))
-      .map { filterChanges =>
-        Right(GetFilterChangesResponse(filterChanges))
-      }
-  }
+  //   filterManager
+  //     .askFor[FM.FilterChanges](FM.GetFilterChanges(req.filterId))
+  //     .map { filterChanges =>
+  //       Right(GetFilterChangesResponse(filterChanges))
+  //     }
+  // }
 
-  def getFilterLogs(req: GetFilterLogsRequest): ServiceResponse[GetFilterLogsResponse] = {
-    implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
-    filterManager
-      .askFor[FM.FilterLogs](FM.GetFilterLogs(req.filterId))
-      .map { filterLogs =>
-        Right(GetFilterLogsResponse(filterLogs))
-      }
-  }
+  // def getFilterLogs(req: GetFilterLogsRequest): ServiceResponse[GetFilterLogsResponse] = {
+  //   implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
+  //   filterManager
+  //     .askFor[FM.FilterLogs](FM.GetFilterLogs(req.filterId))
+  //     .map { filterLogs =>
+  //       Right(GetFilterLogsResponse(filterLogs))
+  //     }
+  // }
 
-  def getLogs(req: GetLogsRequest): ServiceResponse[GetLogsResponse] = {
-    implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
-    import req.filter._
+  // def getLogs(req: GetLogsRequest): ServiceResponse[GetLogsResponse] = {
+  //   implicit val timeout: Timeout = Timeout(filterConfig.filterManagerQueryTimeout)
+  //   import req.filter._
 
-    filterManager
-      .askFor[FM.LogFilterLogs](FM.GetLogs(fromBlock, toBlock, address, topics))
-      .map { filterLogs =>
-        Right(GetLogsResponse(filterLogs))
-      }
-  }
+  //   filterManager
+  //     .askFor[FM.LogFilterLogs](FM.GetLogs(fromBlock, toBlock, address, topics))
+  //     .map { filterLogs =>
+  //       Right(GetLogsResponse(filterLogs))
+  //     }
+  // def getUncleCountByBlockNumber(
+  //     req: GetUncleCountByBlockNumberRequest
+  // ): ServiceResponse[GetUncleCountByBlockNumberResponse] = {
+  //   Task {
+  //     resolveBlock(blockchain, ledger, req.block).map { case ResolvedBlock(block, _) =>
+  //       GetUncleCountByBlockNumberResponse(block.body.uncleNodesList.size)
+  //     }
+  //   }
+  // }
+
+  // def getUncleCountByBlockHash(
+  //     req: GetUncleCountByBlockHashRequest
+  // ): ServiceResponse[GetUncleCountByBlockHashResponse] = {
+  //   Task {
+  //     blockchain.getBlockBodyByHash(req.blockHash) match {
+  //       case Some(blockBody) =>
+  //         Right(GetUncleCountByBlockHashResponse(blockBody.uncleNodesList.size))
+  //       case None =>
+  //         Left(
+  //           JsonRpcError.InvalidParams(s"Block with hash ${Hex.toHexString(req.blockHash.toArray[Byte])} not found")
+  //         )
+  //     }
+  //   }
+  // }
+
+  // def getBlockTransactionCountByNumber(
+  //     req: GetBlockTransactionCountByNumberRequest
+  // ): ServiceResponse[GetBlockTransactionCountByNumberResponse] = {
+  //   Task {
+  //     resolveBlock(blockchain, ledger, req.block).map { case ResolvedBlock(block, _) =>
+  //       GetBlockTransactionCountByNumberResponse(block.body.transactionList.size)
+  //     }
+  //   }
+  // }
 
   private def doCall[A](req: CallRequest)(
       f: (SignedTransactionWithSender, BlockHeader, Option[InMemoryWorldStateProxy]) => A
