+++ src/test/scala/io/iohk/ethereum/jsonrpc/EthFilterServiceSpec.scala
@@ -1,0 +1,3 @@
+package io.iohk.ethereum.jsonrpc
+
+class EthFilterServiceSpec {}
